refactor: reroganize api into base-api api/memes
